[core] Move NoOpCancelable into core module
